Move TextLink back to EventDetailsScreen folder;
